🚚 Move event from notification to scenario domain
